Sun, Jul 23, 2023, 12:27 PM +04:30

+++ CS50x/PSET1/credit/credit.c
@@ -16,7 +16,7 @@ int main(void)
 
 long get_card (void)
 {
-    int c;
+    long c;
     do
     {
         c = get_long("Enter your card number?\n");
